create user and autheticate user
