+++ models/user.py
@@ -6,7 +6,6 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer,primary_key=True)
-    name = Column(String)
     email = Column(String,unique=True)
     password = Column(String)
     is_active = Column(Boolean,default=True)
