+++ routers/user.py
@@ -4,6 +4,7 @@ from utils.jwt_manager import create_token
 from schemas.user import User,UserBase,UserCreate
 from config.database import Session
 from services.user import UserService
+from services.auth import Auth
 
 user_router = APIRouter()
 
@@ -18,8 +19,11 @@ def create_user(user:UserCreate):
     return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,content={"message":"User already exists"})
 
 @user_router.post('/login',tags=['Auth'],status_code=status.HTTP_200_OK)
-def login(user:User):
-    if user.email == "platziuser@fake.com" and user.password == "admin123@":
+def login(user:UserCreate):
+    db = Session()
+    result = UserService(db).get_user_by_email(email=user.email)
+    
+    if result and Auth().verify_password(user.password,result.password):
         token:str = create_token(user.dict())
         return JSONResponse(status_code=status.HTTP_200_OK,content=token)
     return JSONResponse(status_code=status.HTTP_401_UNAUTHORIZED,content={"message":"Unauthorized"})
